first pass of reverting config move
